+++ tests/testthat/test-helpers.R
@@ -7,7 +7,7 @@ test_that("lonlat cartesian conversions", {
   "convert back and forth"
   expect_equal(
     do.call(cartesian_lonlat, lonlat_cartesian(lon, lat)),
-    list(x = lon, y = lat)
+    list(X = lon, Y = lat)
   )
 })
 
@@ -19,19 +19,22 @@ test_that("x checks", {
   expect_error(x_checks(sf::st_cast(x, "LINESTRING"), "x", c("POINT", "POLYGON", "MULTIPOINT", "MULTIPOLYGON")))
 
   "no CRS"
-  x_noCRS <- x
-  sf::st_crs(x_noCRS) <- NA
+  x_noCRS <- sf::st_set_crs(x, NA)
   expect_error(x_checks(x_noCRS, "x", c("POINT", "POLYGON", "MULTIPOINT", "MULTIPOLYGON")))
 
-  "pass"
-  expect_equal(x_checks(x, "x", c("POINT", "POLYGON", "MULTIPOINT", "MULTIPOLYGON")), sf::st_centroid(x))
+  "no geometry"
+  x_emptyGeom <- sf::st_as_sf(
+    data.frame(x = c(1, NA), y = c(1, NA)), 
+    coords = c("x", "y"), 
+    na.fail = FALSE
+  )
+  expect_error(x_checks(x_emptyGeom, "x", c("POINT", "POLYGON", "MULTIPOINT", "MULTIPOLYGON")))
 
+  "pass"
+  expect_no_error(x_checks(x, "x", c("POINT", "POLYGON", "MULTIPOINT", "MULTIPOLYGON")))
 })
 
 test_that("weight checks", {
-  "NULL weight"
-  expect_equal(weight_checks(x, "x", NULL), rep(1, nrow(x)))
-
   "non existent column"
   expect_error(weight_checks(x, "x", "random"))
 
@@ -40,7 +43,7 @@ test_that("weight checks", {
 
   "missing value"
   x_missing_wt <- x
-  x_missing_wt$wts[10] <- NA_real_
+  x_missing_wt$wts[10] <- NA
   expect_error(weight_checks(x_missing_wt, "x", "wts"))
 
   "negative value"
@@ -49,19 +52,33 @@ test_that("weight checks", {
   expect_error(weight_checks(x_neg_wt, "x", "wts"))
 
   "existent column"
-  expect_equal(weight_checks(x, "x", "wts"), x$wts)
+  expect_no_error(weight_checks(x, "x", "wts"))
 })
 
 test_that("group checks", {
-  "NULL group"
-  expect_equal(group_checks(x, "x", NULL), rep("1", nrow(x)))
-
   "non existent column"
   expect_error(group_checks(x, "x", "random"))
 
-  "non numeric column"
-  expect_error(group_checks(x, "x", "wts"))
+  "missing value"
+  x_missing_grp <- x
+  x_missing_grp$grp[10] <- NA
+  expect_error(weight_checks(x_missing_wt, "x", "wts"))
 
   "existent column"
-  expect_equal(group_checks(x, "x", "grp"), x$grp)
+  expect_no_error(group_checks(x, "x", "grp"))
 })
+
+test_that("output", {
+  "tibble"
+  x_tibble <- tibble::tibble(x) |> sf::st_as_sf()
+  expect_s3_class(mean_center(x_tibble), "tbl_df")
+
+  "data.table"
+  x_dt <- data.table::data.table(x) |> sf::st_as_sf()
+  expect_s3_class(mean_center(x_dt), "data.table")
+
+  "crs"
+  expect_equal(sf::st_crs(mean_center(x)), sf::st_crs(x))
+})
+
+