+++ NEWS.md
@@ -1,3 +1,11 @@
-# centr (development version)
+# version 0.2
 
-* Added a `NEWS.md` file to track changes to the package.
+* Warning given if empty geometries returned due to zero total weight groups
+* Groups can be more than just characters
+* Support for `data.table` `sf` objects tested
+* Vignette added for a walkthrough of typical usage 
+
+# version 0.1
+
+* Inital release
+* Published to CRAN